+++ src/Sources/RequestParameter.php
@@ -1,16 +1,13 @@
 <?php
 
-namespace Spatie\AnalyticsTracker\Sources;
+namespace Adzbuck\LaravelUTM\Sources;
 
 use Illuminate\Http\Request;
 
 class RequestParameter
 {
-    protected Request $request;
-
-    public function __construct(Request $request)
+    public function __construct(protected Request $request)
     {
-        $this->request = $request;
     }
 
     public function get(string $key): ?string