+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-All notable changes to `laravel-utm-forwarder` will be documented in this file
+All notable changes to `laravel-utm` will be documented in this file
+
+## 2.0.0 - 2023-04-14
+
+- Fork created
+- Updated dependencies
+- Added last touch tracking
+- Added current
 
 ## 1.0.0 - 202X-XX-XX